+++ test/libsolidity/SolidityExpressionCompiler.cpp
@@ -152,7 +152,11 @@ bytes compileFirstExpression(
 					parametersSize--
 				);
 
-			ExpressionCompiler(context, dev::test::Options::get().optimize).compile(*extractor.expression());
+			ExpressionCompiler(
+				context,
+				RevertStrings::Default,
+				dev::test::Options::get().optimize
+			).compile(*extractor.expression());
 
 			for (vector<string> const& function: _functions)
 				context << context.functionEntryLabel(dynamic_cast<FunctionDefinition const&>(